pyproject.toml
[project]
name = "agentic-job-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "docx2txt>=0.9",
    "fastapi>=0.131.0",
    "gradio>=6.6.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.9",
    "loguru>=0.7.3",
    "pypdf>=6.7.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "sentence-transformers>=5.2.3",
    "tqdm>=4.67.3",
    "unstructured>=0.21.2",
    "uvicorn>=0.41.0",
]
